I fixed a few bugs, the biggest being that operations where not "clearing" the previous value before taking the second input.  IE 1 + 1 = 2 was actually doing 1 + 11 = 12, because it was appending the next digit to the second parameter.
I also added a more "beefy" clear function so it can remove odd states you can get
the system in by clicking an operation but then clear so it is sort of in the middle
of doing a plus or something.  I attempted to add this to a long click event for
the clear button but I was unsucessful so I just upgraded the normal click.